On mastodon - with twtxt generated from it
With twtxt.net disabling AP support, I decided not to wait any longer
and make the jump into ciberlandia.pt. This way I can still have
the sort of ownership and control twtxt provides, with the ease of use
and flexibility a twtxt service provider enables... and a way to
interact with AP accounts, all in one.